Desinstalacion de D3.js y Canvas
