+++ package.json
@@ -8,9 +8,7 @@
     "@testing-library/user-event": "^13.5.0",
     "bootstrap": "^5.2.3",
     "bootswatch": "^5.2.3",
-    "canvas": "^2.11.2",
     "d3": "^7.8.4",
-    "d3-cloud": "^1.2.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.10.0",
